Move rustc version parser functions to new module
