+++ src/utils/mod.rs
@@ -6,6 +6,7 @@ pub use self::copy::{copy_dir, copy_doc_dir};
 pub use self::github_updater::github_updater;
 pub use self::release_activity_updater::update_release_activity;
 pub use self::daemon::start_daemon;
+pub use self::rustc::{parse_rustc_version, get_current_versions, command_result};
 
 mod github_updater;
 mod build_doc;
@@ -13,3 +14,4 @@ mod copy;
 mod release_activity_updater;
 mod daemon;
 mod pubsubhubbub;
+mod rustc;
